refactored the project and (unintentionally) fixed malformet packet error
