+++ raft/raft.go
@@ -14,17 +14,17 @@ import (
 
 
 
-type node struct {
+type Node struct {
 	config   *config.Config
-	raftNode *raft.Raft
-	fsm      *Fsm
+	RaftNode *raft.Raft
+	Fsm      *Fsm
 	logger   *log.Logger
 }
 
 /*
 creates and returns a new node
 */
-func NewNode(config *config.Config, logger *log.Logger) (*node, error){
+func NewNode(config *config.Config, logger *log.Logger) (*Node, error){
 	raftConfig := raft.DefaultConfig()
 	raftConfig.LocalID = raft.ServerID(config.RaftAddress.String())
 	//raftConfig.Logger = log.New(Logger, "", 0)
@@ -69,11 +69,11 @@ func NewNode(config *config.Config, logger *log.Logger) (*node, error){
 		raftNode.BootstrapCluster(configuration)
 		logger.Print("bootstrapping cluster")
 	}
-	return &node{
+	return &Node{
 		config:   config,
-		raftNode: raftNode,
-		logger:      logger,
-		fsm:      fsm,
+		RaftNode: raftNode,
+		logger:   logger,
+		Fsm:      fsm,
 	}, nil
 }
 
@@ -81,7 +81,7 @@ func NewNode(config *config.Config, logger *log.Logger) (*node, error){
 Creates a new node but without persistent storage
 only for tests
  */
-func NewInMemNodeForTesting(config *config.Config, logger *log.Logger) (*node, error){
+func NewInMemNodeForTesting(config *config.Config, logger *log.Logger) (*Node, error){
 
 	raftConfig := raft.DefaultConfig()
 	raftConfig.LocalID = raft.ServerID(config.RaftAddress.String())
@@ -117,11 +117,11 @@ func NewInMemNodeForTesting(config *config.Config, logger *log.Logger) (*node, e
 		raftNode.BootstrapCluster(configuration)
 		logger.Print("bootstrapping cluster")
 	}
-	return &node{
+	return &Node{
 		config:   config,
-		raftNode: raftNode,
-		logger:      logger,
-		fsm:      fsm,
+		RaftNode: raftNode,
+		logger:   logger,
+		Fsm:      fsm,
 	}, nil
 }
 /*
