+++ storage/definitions.go
@@ -10,3 +10,20 @@ type value = string
 //dns
 type hostname = string
 type record = dns.DNSResourceRecord
+
+type Metadata struct {
+	Dnsormetadata bool
+	Service       string
+	Ip            string
+	Type          string
+	Key           string
+	Value         string
+}
+type Dnsresource struct {
+	Dnsormetadata  bool
+	Hostname       string
+	Ip             string
+	RequestType    string
+	ResourceRecord dns.DNSResourceRecord
+}
+
